+++ library/src/actions/partialCheck/utils/_isPartiallyTyped/_isPartiallyTyped.test.ts
@@ -131,7 +131,6 @@ describe('_isPartiallyTyped', () => {
       const dataset: TypedDataset<Input, never> = {
         typed: true,
         value: input,
-        issues: undefined,
       };
       expect(_isPartiallyTyped(dataset, pathList)).toBe(true);
     });