+++ library/src/schemas/looseObject/looseObjectAsync.test.ts
@@ -153,14 +153,25 @@ describe('looseObjectAsync', () => {
     test('for optional entry', async () => {
       await expectNoSchemaIssueAsync(
         looseObjectAsync({ key: optional(string()) }),
-        [{}, { key: undefined }, { key: 'foo' }]
+        [
+          {},
+          // @ts-expect-error
+          { key: undefined },
+          { key: 'foo' },
+        ]
       );
     });
 
     test('for nullish entry', async () => {
       await expectNoSchemaIssueAsync(
         looseObjectAsync({ key: nullish(number()) }),
-        [{}, { key: undefined }, { key: null }, { key: 123 }]
+        [
+          {},
+          // @ts-expect-error
+          { key: undefined },
+          { key: null },
+          { key: 123 },
+        ]
       );
     });