+++ library/src/schemas/object/objectAsync.test-d.ts
@@ -52,8 +52,8 @@ describe('objectAsync', () => {
     test('of input', () => {
       expectTypeOf<InferInput<Schema>>().toEqualTypeOf<{
         key1: string;
-        key2?: string | undefined;
-        key3?: string | null | undefined;
+        key2?: string;
+        key3?: string | null;
         key4: { key: number };
         key5: string;
       }>();
@@ -63,7 +63,7 @@ describe('objectAsync', () => {
       expectTypeOf<InferOutput<Schema>>().toEqualTypeOf<{
         key1: string;
         key2: string;
-        key3?: string | null | undefined;
+        key3?: string | null;
         key4: { key: number };
         readonly key5: string;
       }>();