+++ library/src/schemas/objectWithRest/objectWithRest.test-d.ts
@@ -64,8 +64,8 @@ describe('objectWithRest', () => {
       expectTypeOf<InferInput<Schema>>().toEqualTypeOf<
         {
           key1: string;
-          key2?: string | undefined;
-          key3?: string | null | undefined;
+          key2?: string;
+          key3?: string | null;
           key4: { key: number };
           key5: string;
         } & { [key: string]: boolean }
@@ -77,7 +77,7 @@ describe('objectWithRest', () => {
         {
           key1: string;
           key2: string;
-          key3?: string | null | undefined;
+          key3?: string | null;
           key4: { key: number };
           readonly key5: string;
         } & { [key: string]: boolean }