+++ library/src/schemas/strictObject/strictObject.test-d.ts
@@ -53,8 +53,8 @@ describe('strictObject', () => {
     test('of input', () => {
       expectTypeOf<InferInput<Schema>>().toEqualTypeOf<{
         key1: string;
-        key2?: string | undefined;
-        key3?: string | null | undefined;
+        key2?: string;
+        key3?: string | null;
         key4: { key: number };
         key5: string;
       }>();
@@ -64,7 +64,7 @@ describe('strictObject', () => {
       expectTypeOf<InferOutput<Schema>>().toEqualTypeOf<{
         key1: string;
         key2: string;
-        key3?: string | null | undefined;
+        key3?: string | null;
         key4: { key: number };
         readonly key5: string;
       }>();