+++ library/src/schemas/strictObject/strictObjectAsync.test-d.ts
@@ -58,8 +58,8 @@ describe('strictObjectAsync', () => {
     test('of input', () => {
       expectTypeOf<InferInput<Schema>>().toEqualTypeOf<{
         key1: string;
-        key2?: string | undefined;
-        key3?: string | null | undefined;
+        key2?: string;
+        key3?: string | null;
         key4: { key: number };
         key5: string;
       }>();
@@ -69,7 +69,7 @@ describe('strictObjectAsync', () => {
       expectTypeOf<InferOutput<Schema>>().toEqualTypeOf<{
         key1: string;
         key2: string;
-        key3?: string | null | undefined;
+        key3?: string | null;
         key4: { key: number };
         readonly key5: string;
       }>();